(mondrian) GenILoss fix

+++ mondrian/evaluate.cpp
@@ -38,7 +38,7 @@ vector<vector<vector<string>>> mondrian(vector<vector<string>> dataset,
 								map<int, vector<vector<string>>> hierarchies,
 								vector<int> allQids, vector<int> isQidCat,
 								vector<int> confAtts, const int K,
-								const int L, const int P) {	
+								const int L, const long double P) {	
 	// 1. Create a hierarchy tree for every qid
 	map<int, Tree> trees;
 	// 2. Initialize default generalizations
